Extract SetAuthorizationStatement class
