+++ presto-parser/src/main/java/io/prestosql/sql/tree/SetTableAuthorization.java
@@ -13,83 +13,22 @@
  */
 package io.prestosql.sql.tree;
 
-import com.google.common.collect.ImmutableList;
-
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static java.util.Objects.requireNonNull;
-
 public final class SetTableAuthorization
-        extends Statement
+        extends SetAuthorizationStatement
 {
-    private final QualifiedName source;
-    private final PrincipalSpecification principal;
-
     public SetTableAuthorization(QualifiedName source, PrincipalSpecification principal)
     {
-        super(Optional.empty());
-        this.source = requireNonNull(source, "source is null");
-        this.principal = requireNonNull(principal, "principal is null");
+        super(source, principal);
     }
 
     public SetTableAuthorization(NodeLocation location, QualifiedName source, PrincipalSpecification principal)
     {
-        super(Optional.of(location));
-        this.source = requireNonNull(source, "source is null");
-        this.principal = requireNonNull(principal, "principal is null");
-    }
-
-    public QualifiedName getSource()
-    {
-        return source;
-    }
-
-    public PrincipalSpecification getPrincipal()
-    {
-        return principal;
+        super(location, source, principal);
     }
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context)
     {
         return visitor.visitSetTableAuthorization(this, context);
-    }
-
-    @Override
-    public List<Node> getChildren()
-    {
-        return ImmutableList.of();
-    }
-
-    @Override
-    public int hashCode()
-    {
-        return Objects.hash(source, principal);
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if (this == obj) {
-            return true;
-        }
-        if ((obj == null) || (getClass() != obj.getClass())) {
-            return false;
-        }
-        SetTableAuthorization o = (SetTableAuthorization) obj;
-        return Objects.equals(source, o.source) &&
-                Objects.equals(principal, o.principal);
-    }
-
-    @Override
-    public String toString()
-    {
-        return toStringHelper(this)
-                .add("source", source)
-                .add("principal", principal)
-                .toString();
     }
 }
